starter/requirements.txt
alembic==1.4.3
astroid==2.2.5
Authlib==0.15.2
autopep8==1.5.4
Babel==2.9.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
Click==7.0
cryptography==3.3.1
ecdsa==0.13.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
future==0.17.1
gunicorn==20.0.4
idna==2.10
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.3.1
pylint==2.3.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-jose-cryptodome==1.3.2
pytz==2020.5
requests==2.25.1
six==1.12.0
SQLAlchemy==1.3.3
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.11.1
